refactor(wework): move device git sync to cloud settings
