+++ wework/src/components/settings/MobileSettingsPage.test.tsx
@@ -11,6 +11,17 @@ vi.mock('@/features/model-settings/localCodexSettings', () => ({
   saveLocalCodexPersonality: vi.fn().mockImplementation(value => Promise.resolve(value)),
 }))
 const experimentalFeatures = vi.hoisted(() => ({ enabled: true }))
+const settingsDeviceApi = vi.hoisted(() => ({
+  getAllDevices: vi.fn(),
+  getGitAccountSyncSummary: vi.fn(),
+  syncGitAccounts: vi.fn(),
+}))
+
+vi.mock('./settings-cloud-api', () => ({
+  createSettingsDeviceApi: () => settingsDeviceApi,
+  createSettingsModelApi: () => ({ listModels: vi.fn().mockResolvedValue({ data: [] }) }),
+  createSettingsRemoteTerminalClientFactory: vi.fn(),
+}))
 
 vi.mock('@/features/experimental-features/useExperimentalFeaturesEnabled', () => ({
   useExperimentalFeaturesEnabled: () => experimentalFeatures.enabled,
@@ -19,6 +30,15 @@ vi.mock('@/features/experimental-features/useExperimentalFeaturesEnabled', () =>
 describe('MobileSettingsPage', () => {
   beforeEach(() => {
     experimentalFeatures.enabled = true
+    settingsDeviceApi.getAllDevices.mockReset()
+    settingsDeviceApi.getGitAccountSyncSummary.mockReset()
+    settingsDeviceApi.syncGitAccounts.mockReset()
+    settingsDeviceApi.getAllDevices.mockResolvedValue([])
+    settingsDeviceApi.getGitAccountSyncSummary.mockResolvedValue({
+      accounts: [],
+      effective_count: 0,
+      duplicate_count: 0,
+    })
   })
 
   test('renders mobile settings actions with plugins navigation', async () => {
@@ -40,6 +60,7 @@ describe('MobileSettingsPage', () => {
     expect(pluginsButton).toHaveTextContent('插件')
     expect(pluginsButton.querySelector('.lucide-plug')).toBeInTheDocument()
     expect(screen.getByTestId('mobile-settings-personal-button')).toHaveTextContent('个人')
+    expect(screen.getByTestId('mobile-settings-connections-button')).toHaveTextContent('云端连接')
     expect(screen.getAllByTestId('mobile-settings-worktrees-button')).toHaveLength(1)
     expect(screen.getByTestId('mobile-settings-worktrees-button')).toHaveTextContent('工作树')
     expect(screen.getByTestId('mobile-settings-worktrees-button')).toHaveClass('min-h-[56px]')
@@ -67,6 +88,23 @@ describe('MobileSettingsPage', () => {
 
     await userEvent.click(screen.getByTestId('mobile-settings-plugins-button'))
     expect(onOpenPlugins).toHaveBeenCalledTimes(1)
+  })
+
+  test('opens cloud connection settings on mobile', async () => {
+    render(
+      <AppearanceProvider>
+        <MobileSettingsPage onBack={vi.fn()} />
+      </AppearanceProvider>
+    )
+
+    await userEvent.click(screen.getByTestId('mobile-settings-connections-button'))
+
+    expect(screen.getByTestId('mobile-connections-settings-page')).toBeInTheDocument()
+    expect(screen.getByRole('heading', { name: '云端连接' })).toBeInTheDocument()
+    expect(await screen.findByTestId('git-device-sync-section')).toBeInTheDocument()
+
+    await userEvent.click(screen.getByTestId('mobile-connections-settings-back-button'))
+    expect(screen.getByTestId('mobile-settings-page')).toBeInTheDocument()
   })
 
   test('hides harness settings while experimental features are off', () => {
