+++ wework/src/components/settings/MobileSettingsPage.tsx
@@ -4,6 +4,7 @@ import {
   ChevronRight,
   FolderGit2,
   GitPullRequest,
+  Globe2,
   Info,
   MessageSquareText,
   Package,
@@ -31,6 +32,7 @@ import { QuickPhrasesSettingsPage } from './QuickPhrasesSettingsPage'
 import { RuntimeSettingsPage } from './RuntimeSettingsPage'
 import { HarnessSettingsPage } from './HarnessSettingsPage'
 import { GitHostingSettingsPage } from './GitHostingSettingsPage'
+import { ConnectionsDeviceSettingsPage } from './ConnectionsSettingsPage'
 import type { WorkbenchServices } from '@/features/workbench/workbenchServices'
 import type { DeviceInfo, RuntimeTaskAddress } from '@/types/api'
 
@@ -64,11 +66,40 @@ export function MobileSettingsPage({
     | 'quick-phrases'
     | 'runtimes'
     | 'plugins'
+    | 'connections'
     | 'git-hosting'
     | 'harnesses'
     | 'worktrees'
     | 'archived-conversations'
   >('menu')
+
+  if (activePage === 'connections') {
+    return (
+      <main
+        data-testid="mobile-connections-settings-page"
+        className="flex h-dvh flex-col overflow-hidden bg-background px-5 pb-[max(18px,env(safe-area-inset-bottom))] pt-[max(18px,env(safe-area-inset-top))] text-text-primary"
+      >
+        <header className="flex shrink-0 items-center justify-between">
+          <button
+            type="button"
+            data-testid="mobile-connections-settings-back-button"
+            onClick={() => setActivePage('menu')}
+            className="flex h-11 min-w-[44px] items-center justify-center rounded-full bg-surface text-text-primary hover:bg-muted"
+            aria-label={t('workbench.settings_back_to_app')}
+          >
+            <ArrowLeft className="h-6 w-6" />
+          </button>
+          <h1 className="text-lg font-semibold">
+            {t('workbench.settings_nav_connections', '云端连接')}
+          </h1>
+          <div className="h-11 min-w-[44px]" />
+        </header>
+        <div className="mt-6 min-h-0 flex-1 overflow-auto">
+          <ConnectionsDeviceSettingsPage showHeader={false} />
+        </div>
+      </main>
+    )
+  }
 
   if (activePage === 'git-hosting') {
     return (
@@ -550,6 +581,18 @@ export function MobileSettingsPage({
             <ChevronRight className="h-5 w-5 shrink-0 text-text-muted" />
           </button>
         )}
+        <button
+          type="button"
+          data-testid="mobile-settings-connections-button"
+          onClick={() => setActivePage('connections')}
+          className="flex min-h-[56px] w-full items-center gap-3 rounded-2xl bg-surface px-4 text-left text-base font-medium text-text-primary hover:bg-muted"
+        >
+          <Globe2 className="h-5 w-5 shrink-0 text-text-secondary" />
+          <span className="min-w-0 flex-1 truncate">
+            {t('workbench.settings_nav_connections', '云端连接')}
+          </span>
+          <ChevronRight className="h-5 w-5 shrink-0 text-text-muted" />
+        </button>
         <button
           type="button"
           data-testid="mobile-settings-personal-button"
