+++ wework/src/i18n/locales/zh-CN/common.json
@@ -2310,7 +2310,7 @@
     "execution_environment_check_update": "检查更新",
     "execution_environment_remove": "删除",
     "git_hosting_settings_title": "代码托管",
-    "git_hosting_settings_description": "检测当前设备上的 GitHub CLI 或 GitLab CLI，并将云端保存的 Git 账户安全同步到你明确选择的设备。",
+    "git_hosting_settings_description": "检测当前设备上的 GitHub CLI 或 GitLab CLI，并配置 PR / MR 状态显示。",
     "git_hosting_status_switch": "显示 PR / MR 状态",
     "git_hosting_status_switch_description": "开启后，Wework 会通过当前设备上的 gh 或 glab 查询当前分支状态。关闭后不会执行这些查询。",
     "git_hosting_cli_refresh": "重新检测",
